feat: requires ts-node to start a project for any config that uses ts or jsx
