+++ package.json
@@ -88,7 +88,6 @@
     "@faceless-ui/window-info": "^2.0.2",
     "@monaco-editor/react": "^4.4.6",
     "@swc/core": "^1.3.24",
-    "@swc/register": "^0.1.10",
     "body-parser": "^1.19.0",
     "bson-objectid": "^2.0.1",
     "compression": "^1.7.4",
@@ -181,6 +180,7 @@
     "swc-minify-webpack-plugin": "^1.0.1",
     "terser-webpack-plugin": "^5.0.3",
     "ts-essentials": "^7.0.1",
+    "ts-node": "^10.9.1",
     "url-loader": "^4.1.1",
     "use-context-selector": "^1.4.1",
     "uuid": "^8.1.0",
@@ -284,7 +284,6 @@
     "shelljs": "^0.8.5",
     "slash": "^3.0.0",
     "start-server-and-test": "^1.14.0",
-    "ts-node": "^10.9.1",
     "typescript": "^4.1.2"
   },
   "files": [
