+++ tsconfig.json
@@ -30,6 +30,9 @@
     "skipLibCheck": true, /* Skip type checking of declaration files. */
     "forceConsistentCasingInFileNames": true, /* Disallow inconsistently-cased references to the same file. */
   },
+  "ts-node": {
+    "swc": true,
+  },
   "include": [
     "src/"
   ],
